Move FPS and toaster into main
